requirements.txt
fastapi
uvicorn
tensorflow==2.16.2
numpy
pillow
python-multipart
pydantic
scikit-learn
matplotlib
seaborn
kagglehub
opencv-python-headless
locust
